+++ packages/plugin-link/web/src/typeMapper-loadable.ts
@@ -1,4 +1,4 @@
-import { EXTERNAL_LINK_TYPE, LINK_TYPE, CUSTOM_LINK_TYPE } from './types';
+import { EXTERNAL_LINK_TYPE, LINK_TYPE, CUSTOM_LINK_TYPE, ANCHOR_TYPE } from './types';
 import loadable from '@loadable/component';
 import type { PluginTypeMapper } from 'wix-rich-content-common';
 
@@ -8,5 +8,6 @@ export const typeMapper: PluginTypeMapper = () => ({
     elementType: 'inline',
   },
   [LINK_TYPE]: { component: loadable(() => import('./LinkViewer')), elementType: 'inline' },
+  [ANCHOR_TYPE]: { component: loadable(() => import('./LinkViewer')), elementType: 'inline' },
   [CUSTOM_LINK_TYPE]: { component: loadable(() => import('./LinkViewer')), elementType: 'inline' },
 });