+++ packages/ricos-content/web/src/normalization/Fixtures/index.ts
@@ -8,7 +8,6 @@ export { default as processedInlineImageContentState } from './processedInlineIm
 export { default as inlineGalleryContentState } from './inlineGalleryContentState';
 export { default as processedInlineGalleryContentState } from './processedInlineGalleryContentState';
 
-export { default as AnchorInTextContentState } from './AnchorInTextContentState';
 export { default as AnchorInImageContentState } from './AnchorInImageContentState';
 
 export { default as imageGalleryContentState } from './imageGalleryContentState';