+++ packages/ricos-content/web/src/normalization/normalizeInitialState.spec.ts
@@ -8,7 +8,6 @@ import {
   processedInlineImageContentState,
   inlineGalleryContentState,
   processedInlineGalleryContentState,
-  AnchorInTextContentState,
   AnchorInImageContentState,
   imageGalleryContentState,
   videoInitialContentState,
@@ -772,12 +771,6 @@ describe('normalizeInitialState', () => {
   });
 
   describe('anchor normalizer', () => {
-    it('should remove anchor type (text)', () => {
-      const actual = normalizeInitialState(AnchorInTextContentState, {});
-      expect(actual.entityMap['0'].type).toEqual('LINK');
-      expect(actual.entityMap['1'].type).toEqual('LINK');
-    });
-
     it('should remove anchor type (image)', () => {
       const actual = normalizeInitialState(AnchorInImageContentState, {});
       expect(actual.entityMap['0'].data.config.link).toEqual({ anchor: 'cjvg0' });